Remove zip archives (for now).
We had a long discussion at the TAG F2F and out of that came a dislike
of the zip format because it does not deal with MIME types and such.
Furthermore, http://esdiscuss.org/topic/generic-bundling#content-62
suggests this is not really worth it anyway. HTTP2 will save us.