Fixing #118: add Response.prototype.body
This change introduces `body` property of Response class. Note that we don't
introduce `Request.prototype.body` at this moment and hence some properties and
algorithms which were defined in the Body mixin are moved to Request and
Response.

`bodyUsed` property is now based on `IsDisturbed` defined in the streams spec.
Request.prototype.bodyUsed emulates the behavior without using streams.