Update Fetch with shiny, new CSP hooks
Content Security Policy has defined a set of integrations with Fetch
that make a little more sense than the hand-wavey "Whenever the
user agent requests a resource ..." descriptions in CSP1 and CSP2.
Finally.

This patch updates Fetch's placeholders to link to the new algorithms,
and adds a "CSP list" attribute to response objects in order to support
those algorithms' requirements.