Add Response.prototype.redirected and a new security restriction
This fixes #79 and partially addresses https://github.com/slightlyoff/ServiceWorker/issues/737 (the cache API issue remains for service workers). Open-redirectors will be less vulnerable with this change as navigation always uses redirect mode "manual". This also allows you to inspect whether a response was the result of a redirect although developers are generally encouraged to enforce this through the redirect mode setting so you never end up with an untrustworthy response to begin with.