Call out to Referrer Policy to set policy on redirect
Before following a redirect, Fetch should call out to Referrer Policy to
update the referrer policy on the request, if necessary. (In particular,
this applies when the redirect response contains a Referrer-Policy
header.)

This integration is described in
https://w3c.github.io/webappsec-referrer-policy/#integration-with-fetch,
though it needs to be updated per
https://github.com/w3c/webappsec-referrer-policy/issues/59.

Fixes #334.

PR: https://github.com/whatwg/fetch/pull/335