Block additional "script" MIME types
When request's type is "script", this will make sure we also block
MIME types starting with "audio/" or "video/", or MIME types that are
text/csv. Due to browsers not paying any attention to "script" MIME
types this is blocklist-based, but it's better than nothing. (Use
`X-Content-Type-Options: nosniff` to make it safelist-based.)

PR: https://github.com/whatwg/fetch/pull/379