Allow fetch() to outlive document/workers
See https://github.com/w3c/beacon/pull/27 for related discussion. This basically pulls navigator.sendBeacon() functionality into fetch(). There's a new keepalive flag that when set puts a constraint on the request's body's size while also allowing the fetch to not be terminated when the environment in which it was created goes away.

Fixes #124.