Split up media into video, audio and track
This splits up the `media` destination into three distinct destinations:
`video`, `audio` and `track`.

The reason for that is that track is quite different from video and
audio, and at least in Chromium, treating it internally as
a media resource type is problematic.

This change would allow type specific checks for track.