Make fetch() use "same-origin" credentials by default
Most features in the platform have this default and not having this default causes multiple connections in contemporary implementations. It also causes confusion as switching from XMLHttpRequest to fetch() is not as seamless as it could be.

Making this change will also make it easier for <script type=module> to have this default as per discussion in https://github.com/whatwg/html/issues/2557.

Tests: https://github.com/w3c/web-platform-tests/pull/9911.