Change combined values
This marks XMLHttpRequest's approach in setRequestHeader() for combining values as naïve (it doesn't follow the spirit of HTTP, but is technically allowed) and changes "combined value" to more closely match HTTP semantics.

Tests: https://github.com/web-platform-tests/wpt/pull/13471.

Fixes #757. (Addresses #752 to some extent, but leaving that open for a better API.)