Treat resources requested via FTP as binary data.
Downloading resources over FTP is dangerous in itself, as FTP is a non-securable
protocol. But rendering resources as 'text/html' or similar is even more
dangerous for a variety of reasons (explored to some extent in the blink-dev@
thread linked below). This patch forces FTP resources into an
'application/octet-stream' MIME type, which should prevent them from rendering
as HTML in user agents.

https://groups.google.com/a/chromium.org/d/msg/blink-dev/eopgOoY1QLs/e1tIefOxAAAJ

Closes whatwg/html#4178