Add TAO check
This adds a Timing Allow Origin (TAO) check, trying to be as close to a CORS check as possible. The check is added very close to where the CORS check is, in the HTTP fetch algorithm. It is added so that service worker responses are also checked. When the check fails, it is stored in the request (and once it fails, future redirects cannot cause it to pass). This check is similar to https://w3c.github.io/resource-timing/#dfn-timing-allow-check but the tainted bool is replaced with the tainted origin flag to align with CORS.