Remove trailer support
As noted in https://github.com/whatwg/fetch/issues/772#issuecomment-564515727 the current way it is exposed subsets what HTTP supports and therefore does not feel like a good starting point. Both for the internal and public API.

Additionally, the public-facing API has not seen interest from implementers, at least for the past year and a half.