Add `cross-site` to allowed CORP header values.
This patch adds `cross-site` to the list of allowed CORP header values. It's a no-op
at this point, but I've started asking folks to set this header in preparation for COEP
rollout, and one has pointed me to the spec, noting that the value I'm suggesting they
add isn't technically legal.