Clarify cors requests need cors response tainting
This change clarifies a (non-normative) statement that "cors" request
mode is what makes a request into a cors request. In fact, for requests
with "cors" request mode, "cors" response tainting is also necessary in
order for the request to be considered a cors request. So this change
refines the relevant statement to make that clear.

Otherwise, without this change, considering the case of a same-origin
GET request whose mode is "cors", the spec is claiming that same-origin
GET request is a cors request.

But because the spec defines a "cors request" as “an HTTP request that
includes an `Origin` header”, a same-origin GET request cannot in fact
be a cors request — because it doesn’t include an `Origin` header. (And
that’s because for GET requests, the spec requires an `Origin` header to
be appended only if the request’s response tainting is "cors"; but for
same-origin requests, the request’s response tainting will be "basic".)