Define Content-Length parser
As the body concept is refactored for #604, XMLHttpRequest will need to use this algorithm for its events. We also want to require browsers to use this algorithm instead of the one defined by HTTP as part of #1156.

Tests: https://github.com/web-platform-tests/wpt/pull/10548 & https://github.com/web-platform-tests/wpt/pull/27837.