Resource Timing integration
This defines the foundation for the PerformanceResourceTiming API and its properties. Note that the names used here are not a 1:1 match, e.g., post-redirect start time is fetchStart in the API.

Note that Fetch doesn't decide when to create and enqueue PerformanceResourceTiming objects, but rather creates a data structure that holds the different time values, and it's up to the caller to invoke the newly-added finalize and report timing. Fetch also has a new callback, processResponseDone, for when the response errors or EOFs, which lets callers like the fetch() API invoke finalize and report timing at the right time. This change also makes fetch() adopt that.

It's the first phase for addressing https://github.com/w3c/resource-timing/issues/252, which contains a link to the other required steps to address Resource Timing/Navigation Timing/HTML/Fetch integration.

Corresponding Resource Timing PR: https://github.com/w3c/resource-timing/pull/261. (Once that lands there'll be a subsequent change to Fetch to resolve the remaining TODO.)

Co-authored-by: Anne van Kesteren <annevk@annevk.nl>