Editorial: make WebSocket use obtain a connection
On closer inspection "obtain a WebSocket connection" does not appear to do anything that "obtain a connection" does not do, especially now the latter has a dedicated argument.

This brings "obtain a connection" closer to the goal of covering all web platform connections and a being a policy-enforcement point for them.