Expose a mechanism for hiding internal timing info
This is needed for https://github.com/whatwg/html/issues/7104
and later on for https://github.com/whatwg/fetch/issues/1215.

Navigation timing reports the timing info from the HTML spec,
so it needs a mechanism to obfuscate the internals. So far that
obfuscation was internal to fetch and was done upon reporting.