Return a fetch controller from the fetch algorithm
This allows callers to perform certain actions on an ongoing fetch. Initially to abort or terminate it, and later on to conclude it.

This will eventually remove the need of using response as the object that retains timing info, as that creates certain ambiguities.

Corresponding Service Worker PR: https://github.com/w3c/ServiceWorker/pull/1620.

Follow-up: #1403 and #1412.