Use controller for manual redirect navigation
Remove the existing "navigate-redirect fetch" algorithm. Instead, the caller uses the controller to perform the next manual redirect.

This simplifies the flow in the HTML Standard flow, as now there is a single "fetch params" struct for the course of the entire navigation. Up until now some data from fetch params would be lost, like timing info and task desination, because navigation fetches create a brand new "fetch params" struct. Now that data is retained, and the only thing that HTML needs to do is call "process the next redirect" on the fetch controller.

Co-authored-by: Domenic Denicola <d@domenic.me>