Add content type to response body info
This defines the content type field as a minimized version of the response's parsed Content-Type header, to give web developers enough information to distinguish responses, but not enough to make it an additional communication channel.

This feature will be used by Resource Timing as per https://github.com/w3c/resource-timing/pull/341.

Tests: ...