Remove HTTP cache talk in HTTP-network fetch
This was overlooked in https://github.com/whatwg/fetch/commit/cbca2c2f3a37084e336e14348de683f8ffa0fed9 and is already handled by HTTP-network-or-cache fetch.

Also remove the "abort when" wrapper as this can all be done synchronously. If the fetch is canceled it's caught immediately after.

And finally use the tracking-vector feature instead of a note.