Define opaque-response blocking
This is good enough for early review, but there are a number of issues that still need resolving: https://github.com/annevk/orb/labels/mvp.

There are also some inline TODO comments.

A PR against HTML is needed to ensure it passes the appropriate metadata for media element and classic script requests. We might also want to depend on HTML for parsing JavaScript.

Add the `validate a partial response` algorithm and the `extract content-range values` algorithm

Update the ORB spec to align with Firefox's implementation

Fix the warnings

Try to use dfn-type=http-header to fix the build error