Introduce `get the MIME type` for request and response
Previously, the spec didn't clarify how to get the MIME type for both
request and response. Thus, this patch will introduce the "get the MIME
type" algorithm to clarify how to get it; then, other places which
requires MIME type can reuse this algorithm.

Also, with moving to use the algorithm, it will fix a potential
null-pointer dereference issue in the `formData()` method because a
request or a response's MIME type could potentially be null.