Allow WebDriver BiDi to override the cache mode
The primary use case here is allowing WebDriver to enforce that requests do not
use the network cache, including the CORS preflight cache. But for
future extension the this works by calling into WebDriver BiDi to get
a cache behavior for the request, which is then translated into
specific implementation on the fetch side (e.g. setting the request's
cache mode, or not using the preflight cache).

For now the only possible values of the cache behavior are "default"
and "bypass", so we only need to check for the latter value.