Allow HTTP scheme fetches to make CORS preflight
This allows navigations to make CORS preflight requests, if the
navigation contains a request that is not safelisted. Navigation does
not (yet) provide APIs for non-safelisted requests; this update to the
fetch spec is a prerequisite for HTML spec changes that might add those
APIs (i.e. PUT method support in forms).