Update HTTPFetch algorithm to support new SW Handle Fetch returns
This will update the HttpFetch algorithm, particularly the handling of
the response for Service Worker fetch (`handle fetch` step). Currently,
the response of handle fetch step is assumed to return the `response`
type (or null, if the ServiceWorker couldn't handle the fetch and need
to fallback to the network request). However, we have changed the
`handle fetch` step to also return `service worker timing info` when the
ServiceWorker static routing API used, so that the corresponding timing
information are correctly exposed when the ServiceWorker could not handle
the fetch. To support this new return type, we need to update the handling
of the response of `handle fetch`.

To expose the Service Worker Timing Info to the resource timing API, we
also associate them to the Fetch Timing Info so that it could be later
referenced.