Align with changes to Service Worker's handle fetch
Update the HTTP fetch algorithm, particularly the handling of the response to Service Worker's handle fetch. Currently, handle fetch is assumed to return null or a response. However, handle fetch can now return service worker timing info when the Service Worker static routing API is used, so that the corresponding timing information are correctly exposed when the ServiceWorker could not handle the fetch (i.e., would previously return null). To support this new return type, we need to update the handling of the response of handle fetch.

To expose the Service Worker timing Info to the resource timing API, we also associate them to the fetch timing info.

Co-authored-by: Anne van Kesteren <annevk@annevk.nl>